adjusted file organization
