+++ tsconfig.json
@@ -1,6 +1,7 @@
 {
     "compilerOptions": {
       "target": "es2015",
-      "module": "es2015"
+      "module": "es2015",
+      "jsx": "react"
     }
 }
